+++ script/deploy/single/07_TellerSetup.s.sol
@@ -1,17 +1,17 @@
 // SPDX-License-Identifier: Apache-2.0
 pragma solidity 0.8.21;
 
-import {RolesAuthority} from "@solmate/auth/authorities/RolesAuthority.sol";
-import {ManagerWithMerkleVerification} from "./../../../src/base/Roles/ManagerWithMerkleVerification.sol";
-import {BoringVault} from "./../../../src/base/BoringVault.sol";
-import {TellerWithMultiAssetSupport} from "./../../../src/base/Roles/TellerWithMultiAssetSupport.sol";
-import {AccountantWithRateProviders} from "./../../../src/base/Roles/AccountantWithRateProviders.sol";
-import {BaseScript} from "../../Base.s.sol";
-import {ConfigReader} from "../../ConfigReader.s.sol";
-import {CrossChainTellerBase} from "../../../src/base/Roles/CrossChain/CrossChainTellerBase.sol";
-import {ERC20} from "@solmate/tokens/ERC20.sol";
-import {Strings} from "@openzeppelin/contracts/utils/Strings.sol";
-import {stdJson as StdJson} from "@forge-std/StdJson.sol";
+import { RolesAuthority } from "@solmate/auth/authorities/RolesAuthority.sol";
+import { ManagerWithMerkleVerification } from "./../../../src/base/Roles/ManagerWithMerkleVerification.sol";
+import { BoringVault } from "./../../../src/base/BoringVault.sol";
+import { TellerWithMultiAssetSupport } from "./../../../src/base/Roles/TellerWithMultiAssetSupport.sol";
+import { AccountantWithRateProviders } from "./../../../src/base/Roles/AccountantWithRateProviders.sol";
+import { BaseScript } from "../../Base.s.sol";
+import { ConfigReader } from "../../ConfigReader.s.sol";
+import { CrossChainTellerBase } from "../../../src/base/Roles/CrossChain/CrossChainTellerBase.sol";
+import { ERC20 } from "@solmate/tokens/ERC20.sol";
+import { Strings } from "@openzeppelin/contracts/utils/Strings.sol";
+import { stdJson as StdJson } from "@forge-std/StdJson.sol";
 
 contract TellerSetup is BaseScript {
     using Strings for address;
@@ -21,19 +21,21 @@ contract TellerSetup is BaseScript {
         deploy(getConfig());
     }
 
-    function deploy(ConfigReader.Config memory config) public virtual override broadcast returns(address){
+    function deploy(ConfigReader.Config memory config) public virtual override broadcast returns (address) {
         TellerWithMultiAssetSupport teller = TellerWithMultiAssetSupport(config.teller);
 
         // add the base asset by default for all configurations
         teller.addAsset(ERC20(config.base));
 
         // add the remaining assets specified in the assets array of config
-        for(uint i; i < config.assets.length; ++i){
+        for (uint256 i; i < config.assets.length; ++i) {
             // add asset
             teller.addAsset(ERC20(config.assets[i]));
 
             // set the cooresponding rate provider
-            string memory key = string(abi.encodePacked(".assetToRateProviderAndPriceFeed.",config.assets[i].toHexString(),".rateProvider"));
+            string memory key = string(
+                abi.encodePacked(".assetToRateProviderAndPriceFeed.", config.assets[i].toHexString(), ".rateProvider")
+            );
             address rateProvider = getChainConfigFile().readAddress(key);
             AccountantWithRateProviders(teller.accountant()).setRateProviderData(ERC20(config.base), true, rateProvider);
         }
