+++ src/base/DecodersAndSanitizers/Protocols/ITB/common/WithdrawableDecoderAndSanitizer.sol
@@ -4,7 +4,7 @@ pragma solidity ^0.8.0;
 /// @title Decoder and sanitizer for Withdrawable
 /// @author IntoTheBlock Corp
 abstract contract WithdrawableDecoderAndSanitizer {
-    function withdraw(address _asset_address, uint) external pure returns (bytes memory addressesFound) {
+    function withdraw(address _asset_address, uint256) external pure returns (bytes memory addressesFound) {
         addressesFound = abi.encodePacked(_asset_address);
     }
